backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
psutil==5.9.6
sentry-sdk[fastapi]==1.38.0
python-json-logger==2.0.7
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pytest==7.4.3
pytest-asyncio==0.21.1